% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-casette.R
\name{as.cassette}
\alias{as.cassette}
\alias{as.cassettepath}
\title{Coerce names, etc. to cassettes}
\usage{
as.cassette(x)

as.cassettepath(x)
}
\arguments{
\item{x}{Input}
}
\value{
a casette of class \code{Cassette}
}
\description{
Coerce names, etc. to cassettes

Coerce to a cassette path
}
\examples{
vcr_configure(dir = "~/fixtures/vcr_cassettes")
insert_cassette("foobar")
as.cassette("foobar")
eject_cassette()
}
