% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_block.R
\name{call_block}
\alias{call_block}
\title{Record a call to a block of code}
\usage{
call_block(cassette, block)
}
\description{
Record a call to a block of code
}
\examples{
\dontrun{
cassette <- cassettes()[[3]]
call_block(cassette, {
  GET("http://httpbin.org/get")
})
}
}

