% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassettes.R
\name{cassettes}
\alias{cassette_current}
\alias{cassette_path}
\alias{cassettes}
\title{List cassettes, get current cassette, etc.}
\usage{
cassettes()

cassette_current()

cassette_path()
}
\description{
List cassettes, get current cassette, etc.
}
\examples{
\dontrun{
# list all cassettes
cassettes()

# list the currently active cassette
cassette_current()

# list the path to cassettes
cassette_path()
}
}

