% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cassette_names.R
\name{check_cassette_names}
\alias{check_cassette_names}
\title{Check cassette names}
\usage{
check_cassette_names(
  pattern = "test-",
  behavior = "stop",
  allowed_duplicates = NULL
)
}
\arguments{
\item{pattern}{(character) regex pattern for file paths to check.
This is done inside of \verb{tests/testthat/}. Default: "test-".}

\item{behavior}{(character) "stop" (default) or "warning". If "warning",
we use \code{immediate.=TRUE} so the warning happens at the top of your
tests rather than you seeing it after tests have run (as would happen
by default).}

\item{allowed_duplicates}{(character) Cassette names that can be duplicated.}
}
\description{
This function is meant to be run during your tests, from a
\href{https://testthat.r-lib.org/reference/test_dir.html#special-files}{\verb{helper-*.R} file}
inside the \code{tests/testthat} directory.

It only checks that cassette names are not duplicated. Note that if you do
need to have duplicated cassette names you can do so by using the
\code{allowed_duplicates} parameter in \code{check_cassette_names()}.
}
