% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcr-package.R
\docType{data}
\name{crul_request}
\alias{crul_request}
\title{An HTTP request as prepared by the \pkg{crul} package}
\format{
A list
}
\description{
The object is a list, and is the object that is passed on to
\pkg{webmockr} and \pkg{vcr} instead of routing through
\pkg{crul} as normal. Used in examples/tests.
}
\keyword{data}
