% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eject.R
\name{eject}
\alias{eject}
\title{Eject a cassette}
\usage{
eject(cassette, options = list())
}
\arguments{
\item{cassette}{(character) One or more cassettes to eject}

\item{options}{Optional list of options to pass to eject}
}
\description{
Eject a cassette
}
\examples{
\dontrun{
x <- insert_cassette(name = "fartloud")
eject(x)
}
}
