% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eject_cassette.R
\name{eject_cassette}
\alias{eject_cassette}
\title{Eject a cassette}
\usage{
eject_cassette(cassettes, options = list())
}
\arguments{
\item{cassettes}{(character) One or more cassettes to eject}

\item{options}{Optional list of options}
}
\description{
Eject a cassette
}
\examples{
\dontrun{
res <- Cassette$new("foobar")

x <- cassettes()
(cas <- as.cassette(x[[1]]))
as.cassette(cas)
as.cassette(cassettes()[[1]])
as.cassette("foobar")

insert_cassette(name = "fartloud")

use_cassette("foobar", GET("http://google.com"))

eject_cassette("foobar")
}
}

