% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persisters-file.R
\docType{data}
\name{FileSystem}
\alias{FileSystem}
\title{File system persister}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FileSystem
}
\arguments{
\item{file_name}{(character) Just he fiile name, not whole path}

\item{path}{(character) Storage directory for cassettes}

\item{content}{(character) content to record to a cassette}
}
\description{
The only built-in cassette persister. Persists cassettes
to the file system.
}
\examples{
\dontrun{
nn <- list.files(cassette_path())

(yy <- FileSystem$new())
yy$get_cassette(file_name = "file4014931b21b.yml")

(yy <- FileSystem$new(file_name = "file4014931b21b.yml"))
yy$get_cassette()

(yy <- FileSystem$new(file_name = "file4014931b21b.yml"))
yy$set_cassette(content = "hello world!")
}
}
\keyword{datasets}

