% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\docType{data}
\name{Hooks}
\alias{Hooks}
\title{Hooks class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Hooks
}
\description{
Hooks class
}
\details{
Helps define new hooks, hold hooks, and accessors to get and use hooks.
}
\examples{
\dontrun{
(x <- Hooks$new())
x$hooks
x$define_hook(hook_type = "foo", fun = function(x) x ^ 2)
x$hooks$foo(4)
x$clear_hooks()
x$hooks
}
}
\keyword{internal}

