% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_interactions.R
\name{http_interactions}
\alias{http_interactions}
\title{Get the http interactions of the current cassette}
\usage{
http_interactions()
}
\value{
object of class \code{HTTPInteractionList} if there is a current
cassette in use, or \code{NullList} if no cassette in use
}
\description{
Get the http interactions of the current cassette
}
\examples{
\dontrun{
http_interactions()
}
}
