% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_cassette.R
\name{is_recording}
\alias{is_recording}
\alias{is_replaying}
\title{Determine if vcr is recording/replaying}
\usage{
is_recording()

is_replaying()
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
\code{\link[=local_cassette]{local_cassette()}} and \code{use_cassette()} set the \code{VCR_IS_RECORDING}
and \code{VCR_IS_REPLAYING} environment variables to make it easy to determine
vcr state without taking a dependency on vcr. These functions show you
how to use them; we expect you to copy and paste these functions into your
own package
}
