% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers-json.R
\name{JSON}
\alias{JSON}
\title{The JSON serializer}
\description{
class with methods for serializing via \pkg{jsonlite}
}
\examples{
\dontrun{
ww <- JSON$new(path = "stuff3")
ww
ww$file_extension
fun <- ww$serialize()
fun(list(http_interactions = list(response = list(body = "bar"))),
  path = ww$path, bytes = FALSE)
ww$deserialize_path()
ww$deserialize_string(string = '["hey", "hi", "hello"]')
ww$deserialize_string(string = '{"foo": "bar"}')
}
}
\keyword{internal}
\section{Super class}{
\code{\link[vcr:Serializer]{vcr::Serializer}} -> \code{JSON}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{JSON$new()}}
\item \href{#method-serialize}{\code{JSON$serialize()}}
\item \href{#method-deserialize_string}{\code{JSON$deserialize_string()}}
\item \href{#method-deserialize_path}{\code{JSON$deserialize_path()}}
\item \href{#method-clone}{\code{JSON$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{JSON} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSON$new(path = NULL, string = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{(character) full path to the yaml file}

\item{\code{string}}{(character) path string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{JSON} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-serialize"></a>}}
\if{latex}{\out{\hypertarget{method-serialize}{}}}
\subsection{Method \code{serialize()}}{
Serializes the given hash using internal fxn write_json
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSON$serialize(x, path, bytes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(list) the object to serialize}

\item{\code{path}}{(character) the file path}

\item{\code{bytes}}{(logical) whether to preserve exact body bytes or not}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(character) the json string to write to disk
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deserialize_string"></a>}}
\if{latex}{\out{\hypertarget{method-deserialize_string}{}}}
\subsection{Method \code{deserialize_string()}}{
Deserializes the given string using jsonlite::fromJSON
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSON$deserialize_string(string = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{string}}{(character) the YAML string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(list) the deserialized object, an R list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deserialize_path"></a>}}
\if{latex}{\out{\hypertarget{method-deserialize_path}{}}}
\subsection{Method \code{deserialize_path()}}{
Deserializes the given string using jsonlite::fromJSON
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSON$deserialize_path()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(list) the deserialized object, an R list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JSON$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
