% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_interaction_list.R
\docType{data}
\name{NullList}
\alias{NullList}
\title{Null list, an empty HTTPInteractionList object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NullList
}
\description{
Null list, an empty HTTPInteractionList object
}
\keyword{datasets}
