% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{paste_desc}
\alias{paste_desc}
\title{Append value if needed}
\usage{
paste_desc(x, y)
}
\arguments{
\item{x}{Existing DESCRIPTION field value}

\item{y}{Value to append}
}
\value{
A character string
}
\description{
Append value if needed
}
\examples{
paste_desc(NA, 1)
paste_desc(1, 1)
paste_desc(2, 1)
}
