% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persisters.R
\docType{data}
\name{Persisters}
\alias{Persisters}
\alias{persister_fetch}
\title{Keeps track of the cassette persisters in a hash-like object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Persisters

persister_fetch(x = "FileSystem")
}
\description{
Keeps track of the cassette persisters in a hash-like object
}
\details{
There's only one option: \code{FileSystem}
}
\examples{
\dontrun{
(aa <- Persisters$new())
aa$name
aa$persisters
yaml_serializer <- aa$persisters$new()
yaml_serializer
}
}
\keyword{internal}

