% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tape.R
\name{play_tape}
\alias{play_tape}
\title{Play back a recorded tape.}
\usage{
play_tape(expr, tape, method = httr::GET)
}
\arguments{
\item{expr}{call. The expression to evaluate with the recorded tape.}

\item{tape}{character. A path to the tape with data to play back.}

\item{method}{function. The function in httr to replace with the tape.}
}
\description{
Play back a recorded tape.
}
