% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/real_connections.R
\name{real_http_connections_allowed}
\alias{real_http_connections_allowed}
\title{Are real http connections allowed?}
\usage{
real_http_connections_allowed()
}
\value{
boolean, \code{TRUE} if real HTTP requests allowed; \code{FALSE} if not
}
\description{
Are real http connections allowed?
}
\examples{
real_http_connections_allowed()
}
