% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{record}
\alias{record}
\title{Record an httr method.}
\usage{
record(expr, method = httr::GET, file = "vcr.txt", snapshot = FALSE,
  callback = function(r) capture.output(dput(r)))
}
\arguments{
\item{expr}{call. The expression to evaluate and record.}

\item{method}{function. The function in httr to run recording for.}

\item{file}{character. A path to the file where the recording should be written.}

\item{snapshot}{logical. If snapshot is true, an RDS tape of the response will be written.}

\item{callback}{function. A function to process the response prior to recording to a file.}
}
\description{
Record an httr method.
}
