% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register-chunk-hook.R
\name{register_chunk_hook}
\alias{register_chunk_hook}
\title{Use vcr in vignettes}
\usage{
register_chunk_hook(prefix = "", dir = "_vcr", ...)
}
\arguments{
\item{prefix}{An optional prefix for the cassette name so that you only
need to make sure that casssette names are unique within the current
vignette. Default: `""``.}

\item{dir}{Directory where to create the cassette file. Default: `"_vcr"``.}

\item{...}{Other arguments passed on to \code{\link[=insert_cassette]{insert_cassette()}}.}
}
\description{
\code{register_chunk_hook()} registers a \code{use_cassette} hook so that you map
a vcr cassette to a knitr chunk. Call it in your setup chunk:

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{r setup\}
#| include: false
vcr::register_chunk_hook()
```
}\if{html}{\out{</div>}}

Then in a chunk you want to use a cassette, set the \code{cassette} chunk option
to the name of the cassette:

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{r\}
#| cassette: name
req <- httr2::request("http://r-project.org")
resp <- httr2::req_perform(req)
```
}\if{html}{\out{</div>}}
}
