% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_ignorer.R
\docType{data}
\name{RequestIgnorer}
\alias{RequestIgnorer}
\title{Request ignorer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RequestIgnorer
}
\description{
Request ignorer
}
\details{
Hook to handle request ignorers, including:
\itemize{
 \item ignore_request -  will ignore any request for which the given
 function returns \code{TRUE}
 \item ignore_hosts - ignore any named host
 \item ignore_localhost - ignore all localhost values (localhost, 127.0.0.1, 0.0.0.0)
 \item ignore_localhost_value - ignore a specific named localhost
}
}
\examples{
\dontrun{
(x <- RequestIgnorer$new())
x$LOCALHOST_ALIASES
x$ignored_hosts
x$ignore_hosts(hosts = "google.com")
x$ignored_hosts
x$ignore_localhost()
x$ignored_hosts
x$ignore_localhost_value('127.0.0.1')
x$ignored_hosts
}
}
\keyword{internal}

