% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_matcher_registry.R
\docType{data}
\name{RequestMatcherRegistry}
\alias{RequestMatcherRegistry}
\title{RequestMatcherRegistry class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RequestMatcherRegistry
}
\description{
RequestMatcherRegistry class
}
\examples{
\dontrun{
(x <- RequestMatcherRegistry$new())
x$default_matchers
x$registry
}
}
\keyword{datasets}

