% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_class.R
\docType{data}
\name{Response}
\alias{Response}
\title{The response of an HTTPInteraction}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Response
}
\arguments{
\item{status}{the status of the response}

\item{body}{the response body}

\item{http_version}{the HTTP version}

\item{adapter_metadata}{Additional metadata used by a specific VCR adapter.}
}
\description{
The response of an HTTPInteraction
}
\examples{
\dontrun{
url <- "http://httpbin.org/post"
body <- list(foo = "bar")
(res <- httr::POST(url, body = body))
(x <- Response$new(
   c(status_code = res$status_code, http_status(res)),
   res$headers,
   content(res, "text"),
   res$all_headers[[1]]$version))
x$body
x$status
x$headers
x$http_version
x$to_hash()
x$from_hash()
}
}
\keyword{datasets}

