% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers.R
\docType{data}
\name{Serializers}
\alias{Serializers}
\alias{serializer_fetch}
\title{Keeps track of the cassette serializers in a hash-like object
Note: LETS JUST USE YAML, EASY PEASY FOR NOW}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Serializers

serializer_fetch(x = "yaml", name)
}
\description{
Keeps track of the cassette serializers in a hash-like object
Note: LETS JUST USE YAML, EASY PEASY FOR NOW
}
\examples{
(aa <- Serializers$new())
aa$name
aa$serializers
yaml_serializer <- aa$serializers$new()
yaml_serializer
}
\keyword{internal}

