% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip-vcr-off.R
\name{skip_if_vcr_off}
\alias{skip_if_vcr_off}
\title{Skip tests if vcr is off}
\usage{
skip_if_vcr_off()
}
\value{
Nothing, skip test.
}
\description{
Custom testthat skippers to skip tests if vcr is turned off via the
environment variable \code{VCR_TURN_OFF}.
}
\seealso{
lightswitch
}
