% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tape.R
\name{snapshot_tape}
\alias{snapshot_tape}
\title{Snapshot a call to a tape.}
\usage{
snapshot_tape(expr, tape, method = httr::GET)
}
\arguments{
\item{expr}{call. The expression to evaluate and record.}

\item{tape}{character. A path to the file to write.}

\item{method}{function. The function in httr to add recording to.}
}
\description{
Snapshot a call to a tape.
}
