% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_string.R
\name{str_splitter}
\alias{str_splitter}
\title{split string every N characters}
\usage{
str_splitter(str, length)
}
\arguments{
\item{str}{(character) a string}

\item{length}{(integer) number of characters to split by}
}
\description{
split string every N characters
}
\examples{
\dontrun{
str = "XOVEWVJIEWNIGOIWENVOIWEWVWEW"
str_splitter(str, 5)
str_splitter(str, 5L)
}
}
