% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_body.R
\name{take_body}
\alias{take_body}
\title{Extract request body}
\usage{
take_body(x, ...)
}
\description{
Extract request body
}
\details{
S3 methods and the packages they are used for
\itemize{
\item request: \code{httr}
\item list: \code{crul}
\item httr2_request: \code{httr2}
}
}
\keyword{internal}
