% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcr.R
\name{turn_off}
\alias{turn_off}
\title{Turns VCR off, so that it no longer handles every HTTP request.}
\usage{
turn_off(cassette, options = list(), ignore_cassettes = FALSE)
}
\arguments{
\item{options}{List of options}

\item{ignore_cassettes}{(logical) Controls what happens when a cassette is
inserted while VCR is turned off. If \code{TRUE} is passed, the cassette insertion
will be ignored; otherwise an error will be raised.}
}
\description{
Turns VCR off, so that it no longer handles every HTTP request.
}

