% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcr.R
\name{turned_off}
\alias{turned_off}
\title{Turns VCR off for the duration of a block.}
\usage{
turned_off(options = list(), ignore_cassettes = FALSE)
}
\arguments{
\item{options}{List of options}

\item{ignore_cassettes}{(logical) Controls what happens when a cassette is
inserted while VCR is turned off. If \code{TRUE} is passed, the cassette insertion
will be ignored; otherwise an error will be raised.}
}
\description{
Turns VCR off for the duration of a block.
}
\seealso{
turn_off, turn_on, turned_on
}

