% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{vcr_last_error}
\alias{vcr_last_error}
\title{Get full suggestion messages for the last vcr cassette failure}
\usage{
vcr_last_error()
}
\description{
Get full suggestion messages for the last vcr cassette failure
}
\examples{
\dontrun{
# vcr_last_error()
}
}
