% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{use_vcr_examples}
\alias{use_vcr_examples}
\title{Use 'vcr' for examples in your package}
\usage{
use_vcr_examples(path = ".")
}
\arguments{
\item{path}{character path to the package}
}
\value{
Nothing: called for file system side effects.
}
\description{
This functions amends DESCRIPTION to set up vcr's roxygen2 machinery.
Call it once when you're setting a new package.
}
