% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcr-package.R
\docType{package}
\name{vcr}
\alias{vcr}
\alias{vcr-package}
\title{vcr: Record HTTP Calls to Disk}
\description{
vcr: Record HTTP Calls to Disk
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}

