% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{vcr_last_request}
\alias{vcr_last_request}
\alias{vcr_last_response}
\title{Retrieve last vcr request/response}
\usage{
vcr_last_request()

vcr_last_response()
}
\description{
When debugging, it's often useful to see the last request and response
respectively. These functions give you what would have been recorded to disk
or replayed from disk. If the last request wasn't recorded, and there was
no matching request, \code{vcr_last_response} will return \code{NULL}.
}
