% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{vcr_logging}
\alias{vcr_logging}
\alias{vcr_log_file}
\alias{vcr_log_info}
\title{vcr log file setup}
\usage{
vcr_log_file(file, overwrite = TRUE)

vcr_log_info(message, include_date = TRUE)
}
\arguments{
\item{file}{(character) a file path, required}

\item{overwrite}{(logical) whether or not to overwrite the file at
'file' if it already exists. Default: \code{TRUE}}

\item{message}{(character) a message to log}

\item{include_date}{(logical) include date and time in each log entry.
Default: \code{FALSE}}
}
\description{
vcr log file setup
}
\examples{
# user workflow
vcr_configuration()
vcr_configure(log = TRUE)
vcr_configuration()

readLines("vcr.log") # empty

# log messages
vcr_log_info("hello world!")
readLines("vcr.log")
vcr_log_info("foo bar")
readLines("vcr.log")
## many messages
vcr_log_info(c("brown cow", "blue horse"))
readLines("vcr.log")
vcr_log_info(c("brown cow", "blue horse", "green goat"))
readLines("vcr.log")

# standalone workflow
# set a file to log to
vcr_log_file((f <- tempfile()))
readLines(f) # empty

# log messages
vcr_log_info("hello world!")
readLines("vcr.log")
vcr_log_info("foo bar")
readLines("vcr.log")

# cleanup
# unlink(f)
}
\keyword{internal}
