% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcr_test_path.R
\name{vcr_test_path}
\alias{vcr_test_path}
\title{Locate file in tests directory}
\usage{
vcr_test_path(...)
}
\arguments{
\item{...}{Character vectors giving path component.}
}
\value{
A character vector giving the path
}
\description{
This function, similar to \code{testthat::test_path()}, is designed to work both
interactively and during tests, locating files in the \verb{tests/} directory.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
vcr_test_path("fixtures")
\dontshow{\}) # examplesIf}
}
