% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_class.R
\name{VcrResponse}
\alias{VcrResponse}
\title{A response object}
\description{
Custom vcr http response object
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{status}}{the status of the response}

\item{\code{headers}}{the response headers}

\item{\code{body}}{the response body}

\item{\code{disk}}{a boolean}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-VcrResponse-new}{\code{VcrResponse$new()}}
\item \href{#method-VcrResponse-print}{\code{VcrResponse$print()}}
\item \href{#method-VcrResponse-clone}{\code{VcrResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VcrResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-VcrResponse-new}{}}}
\subsection{Method \code{new()}}{
Create a new VcrResponse object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VcrResponse$new(status, headers, body, disk)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{status}}{the status of the response}

\item{\code{headers}}{the response headers}

\item{\code{body}}{the response body}

\item{\code{disk}}{boolean, is body a file on disk}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{VcrResponse} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VcrResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-VcrResponse-print}{}}}
\subsection{Method \code{print()}}{
print method for the \code{VcrResponse} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VcrResponse$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-VcrResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-VcrResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{VcrResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
