% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{write_json_callback}
\alias{write_json_callback}
\title{A callback for writing the output as JSON content.}
\usage{
write_json_callback(r)
}
\arguments{
\item{r.}{The raw response from the httr call.}
}
\description{
A callback for writing the output as JSON content.
}
