% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers-yaml.R
\docType{data}
\name{YAML}
\alias{YAML}
\title{The YAML serializer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
YAML
}
\description{
The YAML serializer
}
\examples{
(yy <- YAML$new())
library("httr")
x <- GET("http://httpbin.org/get")
yy$serialize(x)
yy$deserialize_path()
# yy$deserialize_string() # if passed on instantiation
yy$deserialize_string(string = "- hey\\n- hi\\n- hello")
yy$deserialize_string(string = "- foo\\n- bar\\n- 3.14")
}
\keyword{datasets}

