% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-casettes.R
\name{as.cassette}
\alias{as.cassette}
\alias{as.cassettepath}
\title{Coerce names, etc. to cassettes}
\usage{
as.cassette(x)

as.cassettepath(x)
}
\arguments{
\item{x}{Input}
}
\description{
Coerce names, etc. to cassettes

Coerce to a cassette path
}

