% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassette_class.R
\docType{data}
\name{Cassette}
\alias{Cassette}
\title{Cassette handler}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Cassette
}
\description{
Cassette handler
}
\details{
erb - a logical or list of variables
}
\examples{
\dontrun{
res <- Cassette$new("teddybear")
res <- Cassette$new("foobar", record = "all")
res$file()
res$originally_recorded_at()
res$recording()
res$serializable_hash()
res$eject()
res$should_remove_matching_existing_interactions()
res$storage_key()
}
}
\keyword{datasets}

