% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassettes.R
\name{cassettes}
\alias{cassettes}
\alias{cassette_current}
\alias{cassette_path}
\title{List cassettes, get current cassette, etc.}
\usage{
cassettes(on_disk = TRUE, verb = FALSE)

cassette_current()

cassette_path(def = "~/vcr/vcr_cassettes")
}
\arguments{
\item{on_disk}{(logical) Check for cassettes on disk + cassettes in session
(\code{TRUE}), or check for only cassettes in session (\code{FALSE}). Default: \code{TRUE}}

\item{verb}{(logical) verbose messages}

\item{def}{(character) base directory for cassettes}
}
\description{
List cassettes, get current cassette, etc.
}
\examples{
vcr_configure("~/fixtures/vcr_cassettes")

# list all cassettes
cassettes()
cassettes(on_disk = FALSE)

# list the currently active cassette
insert_cassette("stuffthings")
cassette_current()
eject_cassette()

# list the path to cassettes
cassette_path()
vcr_configure("foo")
cassette_path()
}
