% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cassette_names.R
\name{check_cassette_names}
\alias{check_cassette_names}
\title{Check cassette names}
\usage{
check_cassette_names(pattern = "test-", behavior = "stop")
}
\arguments{
\item{pattern}{(character) regex pattern for file paths to check.
this is done inside of \verb{tests/testthat/}. default: "test-"}

\item{behavior}{(character) "stop" (default) or "warning". if "warning",
we use \code{immediate.=TRUE} so the warning happens at the top of your
tests rather than you seeing it after tests have run (as would happen
by default)}
}
\description{
Check cassette names
}
\details{
This function is meant to be run in a \code{helper-pkgname.R} file
inside the \code{tests/testthat} directory.
}
