% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_interaction.R
\docType{class}
\name{HTTPInteraction}
\alias{HTTPInteraction}
\title{HTTPInteraction class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
HTTPInteraction
}
\description{
HTTPInteraction class
}
\section{Methods}{

\describe{
  \item{\code{request}}{A Request class object}
  \item{\code{response}}{A Response class object}
  \item{\code{recorded_at}}{Time http interaction recorded at}
}
}
\examples{
\dontrun{
# make the request
url <- "http://httpbin.org/post"
body <- list(foo = "bar")
res <- httr::POST(url, body = body)

# request
(request <- Request$new("POST", url, body, res$headers))
# response
(response <- Response$new(
   http_status(res),
   res$headers,
   content(res, "text"),
   res$all_headers[[1]]$version))

(x <- HTTPInteraction$new(request = request, response = response))
x$recorded_at
x$to_hash()
x$from_hash()
}
}
\keyword{datasets}

