% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_interaction.R
\docType{data}
\name{HTTPInteraction}
\alias{HTTPInteraction}
\title{HTTPInteraction class}
\arguments{
\item{request}{A \code{Request} class object}

\item{response}{A \code{VcrResponse} class object}

\item{recorded_at}{Time http interaction recorded at}
}
\description{
HTTPInteraction class
}
\details{
\strong{Methods}
\describe{
\item{\code{to_hash()}}{
Create a hash.
}
\item{\code{from_hash()}}{
Get a hash back to an R list.
}
}
}
\examples{
\dontrun{
# make the request
url <- "https://httpbin.org/post"
body <- list(foo = "bar")
cli <- crul::HttpClient$new(url = url)
res <- cli$post(body = body)

# request
(request <- Request$new("POST", uri = url,
  body = body, headers = res$response_headers))
# response
(response <- VcrResponse$new(
   res$status_http(),
   res$response_headers,
   res$parse("UTF-8"),
   res$response_headers$status))

(x <- HTTPInteraction$new(request = request, response = response))
x$recorded_at
x$to_hash()
x$from_hash()
}
}
\keyword{datasets}
