% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_interaction.R
\docType{data}
\name{HTTPInteraction}
\alias{HTTPInteraction}
\title{HTTPInteraction class}
\arguments{
\item{request}{A Request class object}

\item{response}{A Response class object}

\item{recorded_at}{Time http interaction recorded at}
}
\description{
HTTPInteraction class
}
\details{
\strong{Methods}
  \describe{
    \item{\code{to_hash()}}{
      Create a hash.
    }
    \item{\code{from_hash()}}{
      Get a hash back to an R list.
    }
  }
}
\examples{
\dontrun{
# make the request
url <- "http://httpbin.org/post"
body <- list(foo = "bar")
res <- httr::POST(url, body = body)

# request
(request <- Request$new("POST", url, body, res$headers))
# response
(response <- Response$new(
   http_status(res),
   res$headers,
   content(res, "text"),
   res$all_headers[[1]]$version))

(x <- HTTPInteraction$new(request = request, response = response))
x$recorded_at
x$to_hash()
x$from_hash()
}
}
\keyword{datasets}

