% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_interaction_list.R
\docType{data}
\name{HTTPInteractionList}
\alias{HTTPInteractionList}
\title{HTTPInteractionList class}
\arguments{
\item{interactions}{list of interaction class}

\item{request_matchers}{xxx}

\item{allow_playback_repeats}{xxx}

\item{parent_list}{xxx}

\item{used_interactions}{xxx}

\item{log_prefix}{xxx}

\item{request}{xxx}
}
\description{
HTTPInteractionList class
}
\details{
\strong{Methods}
  \describe{
    \item{\code{response_for()}}{
      xxx.
    }
    \item{\code{has_interaction_matching()}}{
      xx.
    }
    \item{\code{has_used_interaction_matching()}}{
      xxx.
    }
    \item{\code{remaining_unused_interaction_count()}}{
      Number of unused interactions.
    }
    \item{\code{assert_no_unused_interactions()}}{
      xxx.
    }
  }
\strong{Private Methods}
 \describe{
    \item{\code{has_unused_interactions()}}{
      Are there areny unused interactions?
    }
    \item{\code{matching_interaction_index_for()}}{
      xxx.
    }
    \item{\code{matching_used_interaction_for()}}{
      xxxx.
    }
    \item{\code{interaction_matches_request()}}{
      Check if a request matches an interaction.
    }
    \item{\code{from_hash()}}{
      Get a hash back.
    }
    \item{\code{request_summary()}}{
      Get a request summary.
    }
    \item{\code{response_summary()}}{
      Get a response summary.
    }
  }
}
\examples{
\dontrun{
vcr_configure(
 dir = "fixtures/vcr_cassettes",
 record = "once"
)

# make interactions
## make the request
url <- "http://httpbin.org/post"
body <- list(foo = "bar")
res <- httr::POST(url, body = body)

## request
(request <- Request$new("POST", url, body, res$headers))
## response
(response <- Response$new(
   c(status_code = res$status_code, http_status(res)),
   res$headers,
   content(res, "text"),
   res$all_headers[[1]]$version))
## make an interaction
(inter <- HTTPInteraction$new(request = request, response = response))

# make an interactionlist
(x <- HTTPInteractionList$new(
   interactions = list(inter),
   request_matchers = vcr_configuration()$match_requests_on
))
x$request_matchers
x$log_prefix
x$parent_list
x$used_interactions
x$allow_playback_repeats
x$interactions
}
}
\keyword{datasets}

