% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_interaction_list.R
\docType{data}
\name{HTTPInteractionList}
\alias{HTTPInteractionList}
\title{HTTPInteractionList class}
\arguments{
\item{interactions}{list of interaction class}

\item{request_matchers}{xxx}

\item{allow_playback_repeats}{xxx}

\item{parent_list}{xxx}

\item{used_interactions}{xxx}

\item{log_prefix}{xxx}

\item{request}{xxx}
}
\description{
HTTPInteractionList class
}
\details{
\strong{Methods}
\describe{
\item{\code{response_for(request)}}{
xxx.
}
\item{\code{has_interaction_matching(request)}}{
xx.
}
\item{\code{has_used_interaction_matching(request)}}{
xxx.
}
\item{\code{remaining_unused_interaction_count()}}{
Number of unused interactions (numeric)
}
\item{\code{assert_no_unused_interactions()}}{
xxx.
}
}
\strong{Private Methods}
\describe{
\item{\code{has_unused_interactions()}}{
Are there areny unused interactions?
}
\item{\code{matching_interaction_index_for()}}{
xxx.
}
\item{\code{matching_used_interaction_for(request)}}{
xxxx.
}
\item{\code{interaction_matches_request(request, interaction)}}{
Check if a request matches an interaction (logical)
}
\item{\code{from_hash()}}{
Get a hash back.
}
\item{\code{request_summary(z)}}{
Get a request summary (character)
}
\item{\code{response_summary(z)}}{
Get a response summary (character)
}
}
}
\examples{
\dontrun{
vcr_configure(
 dir = "fixtures/vcr_cassettes",
 record = "once"
)

# make interactions
## make the request
### turn off mocking
crul::mock(FALSE)
url <- "https://httpbin.org/post"
cli <- crul::HttpClient$new(url = url)
res <- cli$post(body = list(a = 5))

## request
(request <- Request$new("POST", url, body, res$headers))
## response
(response <- VcrResponse$new(
   res$status_http(),
   res$response_headers,
   res$parse("UTF-8"),
   res$response_headers$status))
## make an interaction
(inter <- HTTPInteraction$new(request = request, response = response))

# make an interactionlist
(x <- HTTPInteractionList$new(
   interactions = list(inter),
   request_matchers = vcr_configuration()$match_requests_on
))
x$request_matchers
x$log_prefix
x$parent_list
x$parent_list$response_for()
x$parent_list$has_interaction_matching()
x$parent_list$has_used_interaction_matching()
x$parent_list$remaining_unused_interaction_count()
x$used_interactions
x$allow_playback_repeats
x$interactions
x$response_for(request)
}
}
\keyword{datasets}
