% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persisters.R
\docType{data}
\name{Persisters}
\alias{Persisters}
\alias{persister_fetch}
\title{Keeps track of the cassette persisters in a hash-like object}
\usage{
persister_fetch(x = "FileSystem", file_name)
}
\arguments{
\item{persisters}{A list}

\item{name}{(character) Persister name, only option right now is "FileSystem"}
}
\description{
Keeps track of the cassette persisters in a hash-like object
}
\details{
There's only one option: \code{FileSystem}

\strong{Private Methods}
\describe{
\item{\code{persister_get()}}{
Gets and sets a named persister
}
}
}
\examples{
(aa <- Persisters$new())
aa$name
aa$persisters
yaml_serializer <- aa$persisters$new()
yaml_serializer
}
\keyword{internal}
