% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_class.R
\docType{data}
\name{Request}
\alias{Request}
\title{The request of an HTTPInteraction}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Request
}
\arguments{
\item{method}{the HTTP method (i.e. :head, :options, :get, :post, :put, :patch or :delete)}

\item{uri}{the request URI}

\item{body}{the request body}

\item{headers}{the request headers}
}
\description{
The request of an HTTPInteraction
}
\details{
\strong{Methods}
  \describe{
    \item{\code{to_hash()}}{
      Create a hash.
    }
    \item{\code{from_hash()}}{
      Get a hash back to an R list.
    }
  }
}
\examples{
\dontrun{
url <- "http://httpbin.org/post"
body <- list(foo = "bar")
res <- httr::POST(url, body = body)
(x <- Request$new("POST", url, body, res$headers))
x$body
x$method
x$uri
x$headers
x$to_hash()
x$from_hash()

vcr_configure(
 dir = "fixtures/vcr_cassettes",
 record = "once",
 preserve_exact_body_bytes_for = TRUE
)
}
}
\keyword{internal}

