% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_handler.R
\name{RequestHandler}
\alias{RequestHandler}
\alias{RequestHandlerHttr2}
\alias{RequestHandlerHttr}
\alias{RequestHandlerCrul}
\title{Request handlers}
\description{
These are internal classes that should not be used by users.
}
\keyword{internal}
