% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_matcher_registry.R
\docType{data}
\name{RequestMatcherRegistry}
\alias{RequestMatcherRegistry}
\title{RequestMatcherRegistry class}
\arguments{
\item{name}{matcher name}

\item{func}{function that describes a matcher, should return
a single boolean}

\item{r1, r2}{two \link{Request} class objects}
}
\description{
RequestMatcherRegistry class
}
\details{
\strong{Methods}
\describe{
\item{\code{register(name, func)}}{
Register a custom matcher.
}
\item{\code{register_built_ins()}}{
Register all built in matchers.
}
\item{\code{try_to_register_body_as_json(r1, r2)}}{
Try to register body as JSON.
}
}
}
\examples{
\dontrun{
(x <- RequestMatcherRegistry$new())
x$default_matchers
x$registry
}
}
\keyword{datasets}
