% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_class.R
\docType{data}
\name{Response}
\alias{Response}
\title{The response of an HTTPInteraction}
\arguments{
\item{status}{the status of the response}

\item{headers}{the response headers}

\item{body}{the response body}

\item{http_version}{the HTTP version}

\item{adapter_metadata}{Additional metadata used by a specific VCR adapter.}
}
\description{
The response of an HTTPInteraction
}
\details{
\strong{Methods}
  \describe{
    \item{\code{to_hash()}}{
      Create a hash.
    }
    \item{\code{from_hash()}}{
      Get a hash back to an R list.
    }
  }
}
\examples{
\dontrun{
url <- "http://httpbin.org/post"
body <- list(foo = "bar")
(res <- httr::POST(url, body = body))
(x <- Response$new(
   c(status_code = res$status_code, httr::http_status(res)),
   res$headers,
   content(res, "text"),
   res$all_headers[[1]]$version))
x$body
x$status
x$headers
x$http_version
x$to_hash()
x$from_hash()
}
}
\keyword{internal}
