% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers.R
\docType{data}
\name{Serializers}
\alias{Serializers}
\alias{serializer_fetch}
\title{Keeps track of the cassette serializers in a hash-like object}
\usage{
serializer_fetch(x = "yaml", name)
}
\arguments{
\item{name}{Name of a serializer. Only option is "yaml" for now}

\item{serializers}{List of serializers}
}
\description{
Keeps track of the cassette serializers in a hash-like object
}
\details{
\strong{Private Methods}
\describe{
\item{\code{serialize_get()}}{
Gets a named serializer. This is also run on initialize of this function
}
}
}
\examples{
\dontrun{
(aa <- Serializers$new())
aa$name
aa$serializers
yaml_serializer <- aa$serializers$new()
yaml_serializer
}
}
\keyword{internal}
