% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R, R/register-chunk-hook.R
\name{setup_knitr}
\alias{setup_knitr}
\title{Use vcr in vignettes}
\usage{
setup_knitr(prefix = NULL, dir = "_vcr", ...)

setup_knitr(prefix = NULL, dir = "_vcr", ...)
}
\arguments{
\item{prefix}{An prefix for the cassette name to make cassettes unique across
vignettes. Defaults to the file name (sans extension) of the currently
executing vignette.}

\item{dir}{Directory where to create the cassette file, relative to
the directory containing the vignette. Default: `"_vcr"``.}

\item{...}{Other arguments passed on to \code{\link[=insert_cassette]{insert_cassette()}}.}
}
\description{
\code{setup_knitr()} registers a \code{use_cassette} hook so that you map
a vcr cassette to a knitr chunk. Call it in your setup chunk:

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{r setup\}
#| include: false
vcr::setup_knitr()
```
}\if{html}{\out{</div>}}

Then in a chunk where you want to use a cassette, set the \code{cassette} chunk option
to the name of the cassette:

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{r\}
#| cassette: name
req <- httr2::request("http://r-project.org")
resp <- httr2::req_perform(req)
```
}\if{html}{\out{</div>}}

\code{setup_knitr()} registers a \code{use_cassette} hook so that you map
a vcr cassette to a knitr chunk. Call it in your setup chunk:

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{r setup\}
#| include: false
vcr::setup_knitr()
```
}\if{html}{\out{</div>}}

Then in a chunk where you want to use a cassette, set the \code{cassette} chunk option
to the name of the cassette:

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{r\}
#| cassette: name
req <- httr2::request("http://r-project.org")
resp <- httr2::req_perform(req)
```
}\if{html}{\out{</div>}}
}
