% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcr_setup.R
\name{use_vcr}
\alias{use_vcr}
\title{Setup vcr for a package}
\usage{
use_vcr(dir = ".", verbose = TRUE)
}
\arguments{
\item{dir}{(character) path to package root. default's to
current directory}

\item{verbose}{(logical) print progress messages. default: \code{TRUE}}
}
\value{
only messages about progress, returns invisible()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{use_vcr()} is deprecated because it is no longer needed. In lifecycle v2,
\code{\link[=local_cassette]{local_cassette()}} and friends automatically use \verb{tests/testthat/_vcr}
without any need for additional configuration.
}
\details{
Sets a minimum vcr version, which is usually the latest
(stable) version on CRAN. You can of course easily remove or change
the version requirement yourself after running this function.
}
\keyword{internal}
