% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcr_setup.R
\name{use_vcr}
\alias{use_vcr}
\title{Setup vcr for a package}
\usage{
use_vcr(dir = ".", verbose = TRUE)
}
\arguments{
\item{dir}{(character) path to package root. default's to
current directory}

\item{verbose}{(logical) print progress messages. default: \code{TRUE}}
}
\value{
only messages about progress, returns invisible()
}
\description{
Setup vcr for a package
}
