% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-example.R
\name{vcr-example}
\alias{vcr-example}
\title{Trying out our custom tag}
\description{
Trying out our custom tag
}
\examples{
\dontshow{vcr::insert_example_cassette('bla', package = 'vcr')}
if (rlang::check_installed("httr2")) {
  httr2::request("http://r-project.org")
}
\dontshow{vcr::eject_cassette()}
}
\keyword{internal}
