% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcr-package.R
\docType{package}
\name{vcr}
\alias{vcr}
\alias{vcr-package}
\alias{vcr-package}
\title{vcr: Record HTTP Calls to Disk}
\description{
\pkg{vcr} records test suite 'HTTP' requests and replay them during
future runs.
}
\section{Backstory}{

A Ruby gem of the same name (\code{vcr}, \url{https://github.com/vcr/vcr}) was
created many years ago and was the original. Ports in many languages
have been done, including this one for R. Check out that github repo
for all the details on how the canonical version works.
}

\section{Main funcitons}{

The \link{use_cassette} function is most likely what you'll want to use. It
sets the cassette you want to record to, inserts the cassette, and then
ejects the cassette, recording the interactions to the cassette.

Instead, you can use \link{insert_cassette}, but then you have to make sure
to use \link{eject_cassette}.
}

\section{vcr configuration}{

\link{vcr_configure} is the function to use to set R session wide settings.
See it's manual file for help.
}

\section{Record modes}{

See \link{recording} for help on record modes.
}

\section{Request matching}{

See \link{request-matching} for help on the many request matching options.
}

\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
