% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcr-package.R
\docType{package}
\name{vcr}
\alias{vcr}
\alias{vcr-package}
\title{vcr: Record HTTP Calls to Disk}
\description{
\pkg{vcr} records test suite 'HTTP' requests and replay them during
future runs.
}
\details{
Check out the \href{https://books.ropensci.org/http-testing/}{http testing book}
for a lot more documentation on \code{vcr}, \code{webmockr}, and \code{crul}
}
\section{Backstory}{

A Ruby gem of the same name (\code{VCR}, \url{https://github.com/vcr/vcr}) was
created many years ago and is the original. Ports in many languages
have been done. Check out that GitHub repo for all the details on
how the canonical version works.
}

\section{Main functions}{

The \link{use_cassette} function is most likely what you'll want to use. It
sets the cassette you want to record to, inserts the cassette, and then
ejects the cassette, recording the interactions to the cassette.

Instead, you can use \link{insert_cassette}, but then you have to make sure
to use \link{eject_cassette}.
}

\section{vcr configuration}{

\link{vcr_configure} is the function to use to set R session wide settings.
See it's manual file for help.
}

\section{Record modes}{

See \link{recording} for help on record modes.
}

\section{Request matching}{

See \link{request-matching} for help on the many request matching options.
}

\author{
Scott Chamberlain
}
