% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcr_test_path.R
\name{vcr_test_path}
\alias{vcr_test_path}
\title{Locate file in tests directory}
\usage{
vcr_test_path(...)
}
\arguments{
\item{...}{Character vectors giving path component. each character string
gets added on to the path, e.g., \code{vcr_test_path("a", "b")} becomes
\code{tests/a/b} relative to the root of the package.}
}
\value{
A character vector giving the path
}
\description{
This function, similar to \code{testthat::test_path()}, is designed to work both
interactively and during tests, locating files in the \verb{tests/} directory.
}
\note{
Beware if you have more than one \code{tests} directories in your package
root. This may not work as intended if that is the case.
}
\examples{
if (interactive()) {
vcr_test_path("fixtures")
}
}
