% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serializers-yaml.R
\docType{data}
\name{YAML}
\alias{YAML}
\title{The YAML serializer}
\arguments{
\item{file_extension}{(character) A file extension}

\item{path}{(character) full path to the yaml file}

\item{string}{(character) path string}
}
\description{
The YAML serializer
}
\details{
\strong{Methods}
\describe{
\item{\code{serialize(x)}}{
Serializes the given hash using internal fxn write_yaml
}
\item{\code{deserialize_string(string = NULL)}}{
Deserializes the given string using yaml::yaml.load
}
\item{\code{deserialize_path()}}{
Deserializes the given string using yaml::yaml.load_file
}
}
}
\examples{
\dontrun{
(yy <- YAML$new())
library("crul")
cli <- crul::HttpClient$new(url = "https://httpbin.org")
x <- cli$get("get")
file <- tempfile()
yy$serialize()(x, file)
yy$deserialize_path()
# yy$deserialize_string() # if passed on instantiation
yy$deserialize_string(string = "- hey\\n- hi\\n- hello")
yy$deserialize_string(string = "- foo\\n- bar\\n- 3.14")
}
}
\keyword{internal}
